App.ts refactoring
